Use SystemNotificationTracker
